+++ app/src/main/java/org/appcenter/inudorm/usecase/DeleteSchedule.kt
@@ -6,7 +6,7 @@ import org.appcenter.inudorm.model.TeamSchedule
 
 class DeleteSchedule : ResultUseCase<Int, Unit>() {
     override suspend fun onExecute(params: Int) {
-        return App.calendarRepository.deleteTeamSchedule(params.toLong())
+        return App.calendarRepository.deleteTeamSchedule(params)
     }
 }
